requirements.txt
streamlit==1.28.1
langchain-google-genai==0.0.5
pdf2image==1.16.3
Pillow==10.0.1
python-dotenv==1.0.0 